ai_service/yolov8-incident-detection/requirements.txt
numpy
opencv-python
torch
torchvision
pillow
scikit-learn
pyyaml
requests
matplotlib
tqdm
albumentations
wandb
pytest